+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added 
+
+- Add annotations for the encryption-provider-operator.
+
 ## [0.7.1] - 2021-12-08
 
 ### Fixed